feat: separated component for pagination
Abstract pagination out to its component to enable proper testing of presenation components and container components
